extract smartify file
