+++ site/html.go
@@ -3,28 +3,9 @@ package site
 import (
 	"bytes"
 	"io"
-	"regexp"
-	"slices"
-	"strings"
 
 	"golang.org/x/net/html"
 )
-
-func Smartify(extension string, contentReader io.Reader) (io.Reader, error) {
-	if extension != ".html" {
-		return contentReader, nil
-	}
-	node, err := html.Parse(contentReader)
-	if err != nil {
-		return nil, err
-	}
-
-	smartifyHTMLNode(node)
-	var buf bytes.Buffer
-	html.Render(&buf, node)
-
-	return &buf, nil
-}
 
 // Find the first p tag in the given html document and return its text content.
 func ExtractFirstParagraph(htmlReader io.Reader) string {
@@ -109,50 +90,4 @@ func getTextContent(node *html.Node) string {
 		textContent += getTextContent(c)
 	}
 	return textContent
-}
-
-func smartifyHTMLNode(node *html.Node) {
-	for node := node.FirstChild; node != nil; node = node.NextSibling {
-		if node.Type == html.ElementNode && slices.Contains(SKIP_TAGS, node.Data) {
-			continue
-		}
-		if node.Type == html.TextNode {
-			node.Data = smartifyString(node.Data)
-		}
-		smartifyHTMLNode(node)
-	}
-}
-
-var SKIP_TAGS = []string{"pre", "code", "kbd", "script", "math"}
-
-var smartifyTransforms = []struct {
-	match *regexp.Regexp
-	repl  string
-}{
-	{regexp.MustCompile("(^|[^[:alnum:]])``(.+?)''"), "$1“$2”"},
-	{regexp.MustCompile(`(^|[^[:alnum:]])'`), "$1‘"},
-	{regexp.MustCompile(`'`), "’"},
-	{regexp.MustCompile(`(^|[^[:alnum:]])"`), "$1“"},
-	{regexp.MustCompile(`"($|[^[:alnum:]])`), "”$1"},
-	{regexp.MustCompile(`(^|\s)--($|\s)`), "$1–$2"},
-	{regexp.MustCompile(`(^|\s)---($|\s)`), "$1—$2"},
-}
-
-var smartifyReplacer *strings.Replacer
-var smartifyReplaceSpans = map[string]string{}
-
-func init() {
-	smartifyReplacer = strings.NewReplacer(
-		"...", "…",
-		"(c)", "©",
-		"(r)", "®",
-		"(tm)", "™",
-	)
-}
-
-func smartifyString(s string) string {
-	for _, rule := range smartifyTransforms {
-		s = rule.match.ReplaceAllString(s, rule.repl)
-	}
-	return smartifyReplacer.Replace(s)
 }
